display all candidates, working on show one candidate

+++ CSIRO/Models/Course.cs
@@ -2,7 +2,7 @@
 {
     public class Course
     {
-        public int CourseID { get; set; }
+        public long CourseID { get; set; }
         public string Title { get; set; }
 
     }
