+++ CSIRO/Models/University.cs
@@ -2,7 +2,7 @@
 {
     public class University
     {
-        public int UniversityID { get; set; }
+        public long UniversityID { get; set; }
         public int Ranking { get; set; }
         public string Name { get; set; }
         public string Country { get; set; }
